# inventory-be/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Automation and Tooling API",
    "description": "Comprehensive API documentation for the structural inventory platform. Includes Auth, RBAC, Users, Locations, Stocks, Requests, Audit, Notification, and Dashboard modules.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Resource not found."
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": { "type": "integer", "example": 100 },
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 10 },
          "totalPages": { "type": "integer", "example": 10 }
        }
      },
...
      "Location": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "Aisle B, Shelf 3" },
          "description": { "type": "string" },
          "floor": { "type": "string", "example": "Ground Floor" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Stock": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "modelNumber": { "type": "string", "example": "SENS-01X" },
          "description": { "type": "string" },
          "brand": { "type": "string", "example": "Siemens" },
          "quantity": { "type": "integer", "example": 50 },
          "uom": { "type": "string", "example": "PCS" },
          "projectType": { "type": "string" },
          "type": { "type": "string", "enum": ["mechanical", "electrical"] },
          "minStockLevel": { "type": "integer", "example": 10 },
          "locationId": { "type": "string", "format": "uuid", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "parameters": {
      "PageParam": {
        "name": "page", "in": "query",
        "schema": { "type": "integer", "default": 1 }
      },
      "LimitParam": {
        "name": "limit", "in": "query",
        "schema": { "type": "integer", "default": 10 }
      },
...
      "IdPathParam": {
        "name": "id", "in": "path", "required": true,
        "schema": { "type": "string", "format": "uuid" }
      }
    }
  },
  "security": [{ "BearerAuth": [] }],
  "paths": {
...
    "/locations/{id}/stocks": {
      "get": {
        "summary": "Get stocks at a location (Requires locations:read)",
        "tags": ["Locations"],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "responses": {
          "200": {
            "description": "Location with its stocks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "location": { "$ref": "#/components/schemas/Location" },
                        "stocks": { "type": "array", "items": { "$ref": "#/components/schemas/Stock" } }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Location not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "getLocationsIdStocks"
      }
    },
    "/public/locations/{id}/stocks": {
      "get": {
        "summary": "Public QR manifest — no auth required",
        "tags": ["Locations"],
        "security": [],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "responses": {
          "200": {
            "description": "Location manifest",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "location": { "$ref": "#/components/schemas/Location" },
                        "stocks": { "type": "array", "items": { "$ref": "#/components/schemas/Stock" } }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Location not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "getPublicLocationsIdStocks"
      }
    },
    "/stocks": {
      "get": {
        "summary": "Get paginated stocks (Requires stocks:read)",
        "tags": ["Stocks"],
        "parameters": [
          { "$ref": "#/components/parameters/PageParam" },
          { "$ref": "#/components/parameters/LimitParam" },
          { "name": "search", "in": "query", "description": "Partial match on modelNumber or brand", "schema": { "type": "string" } },
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["mechanical", "electrical"] } }
        ],
        "responses": {
          "200": {
            "description": "Paginated stocks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Stock" } },
                    "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                  }
                }
              }
            }
          }
        },
        "operationId": "getStocks"
      },
      "post": {
        "summary": "Create singular Stock (Requires stocks:create)",
        "tags": ["Stocks"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Stock" }
            }
          }
        },
        "responses": {
          "201": { "description": "Stock created", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Stock" } } } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "postStocks"
      }
    },
...
    "/stocks/bulk": {
      "post": {
        "summary": "Bulk import stocks from Excel/CSV (Requires stocks:create)",
        "tags": ["Stocks"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["stocks"],
                "properties": {
                  "stocks": { "type": "array", "minItems": 1, "items": { "$ref": "#/components/schemas/Stock" } }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Stocks imported",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "data": { "type": "object", "properties": { "importedCount": { "type": "integer" } } } }
                }
              }
            }
          },
          "400": { "description": "Empty stocks array", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "postStocksBulk"
      }
    },
    "/stocks/{id}": {
      "get": {
        "summary": "Get specific stock (Requires stocks:read)",
        "tags": ["Stocks"],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "responses": {
          "200": { "description": "Stock data", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Stock" } } } } } },
          "404": { "description": "Stock not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "getStocksId"
      },
      "put": {
        "summary": "Partially update stock fields (Requires stocks:update)",
        "tags": ["Stocks"],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Stock" }
            }
          }
        },
        "responses": {
          "200": { "description": "Stock updated", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Stock" } } } } } },
          "404": { "description": "Stock not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "putStocksId"
      },
      "delete": {
        "summary": "Delete Stock (Requires stocks:delete)",
        "tags": ["Stocks"],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "responses": {
          "200": { "description": "Deleted stock returned", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Stock" } } } } } },
          "404": { "description": "Stock not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "deleteStocksId"
      }
    },
    "/stocks/{id}/quantity": {
      "put": {
        "summary": "Manually set stock quantity — writes audit log (Requires stocks:update)",
        "tags": ["Stocks"],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["quantity"],
                "properties": {
                  "quantity": { "type": "integer", "minimum": 0 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Quantity updated — low-stock notification may fire", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Stock" } } } } } },
          "404": { "description": "Stock not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "putStocksIdQuantity"
      }
    },
    "/stocks/{id}/locations": {
      "get": {
        "summary": "Get location distribution for all stocks of same modelNumber (Requires stocks:read)",
        "tags": ["Stocks"],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "responses": {
          "200": {
            "description": "Location distribution array",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "locationId": { "type": "string", "format": "uuid" },
                          "locationName": { "type": "string" },
                          "floor": { "type": "string" },
                          "quantity": { "type": "integer" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Stock not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "getStocksIdLocations"
      }
    },
...
  }
}